airflow/requirements.txt
pandas
snowflake-connector-python
hdfs
pyarrow
snowflake-sqlalchemy